[113] Use Apollo for subscriptions in DiagramWebSocketContainer
Apollo sends an empty (in our case) "extensions" map in the payload,
hence the change in GraphQLPayload on the backend.

We disable the Apollo cache for subscriptions, at least for the
moment. We can not avoid having a `cache` in the ApolloClient, and it
does not seem to cause issues for queries or mutations, but for
subscriptions it creates a lot of warnings in the console. Maybe we
will re-enable the cache later if it makes sense and we understand how
to avoid these warnings, but disable it for now. We did not have one
before, so it should not be an issue.

Compared to the previous version, this one does not handle
'connection_error' at the moment.

Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>